fix reportlist
